requirements.txt
# Existing ML Dependencies
lightfm==1.17
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
psycopg2-binary>=2.9.0

# FastAPI Stack
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Production
gunicorn==21.2.0
